// src/app/core/models/process.model.ts
export interface Process {
    idProcess?: string;
    domainModelDiagram?: string;
    modelingSessions?: string;
    glossaryTerms?: string;
    architecturalPatterns?: string;
    featureIdentification?: string;
    featureBreakdown?: string;
    featurePrioritization?: string;
    featureDescription?: string;
    featureOwnership?: string;
    developmentMilestones?: string;
    resourceAllocation?: string;
    riskManagement?: string;
    classDiagrams?: string;
    interactionDiagrams?: string;
    designReviewNotes?: string;
    designArtifacts?: string;
    implementationTasks?: string;
    codeRepositoryLinks?: string;
    unitTests?: string;
    codeReviewNotes?: string;
    integrationTestingNotes?: string;
    toDo? : string[];      
    inProgress?: string[];      
    done?: string[];      
  }
  